livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.9",
    "langchain-text-splitters>=1.1.0",
    "livekit-agents[hedra,images,langchain,silero,turn-detector]~=1.3",
    "livekit-plugins-cartesia>=1.4.2",
    "livekit-plugins-deepgram>=1.4.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "torch>=2.10.0",
]
